requirements.txt
flask
flask-cors
requests
pandas
sqlalchemy
psycopg2-binary
python-dotenv
gunicorn
geopy
